requirements.txt
django>=4.2.4
django-globus-portal-framework>=0.4.8
globus-sdk>=3.27.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1

# Used only for create_flow.py
gladier